backend/requirements.txt
cryptography
requests
pymongo
grpcio
BeautifulSoup4 
grpcio-tools==1.46.5


